+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Cast
+
+#### Changed
+
+- `verify` command now supports the `--class-hash` for Walnut verification
+
 ## [0.38.0] - 2025-02-25
 
 ### Forge